Tested Adam opt and scaled up gradient free opt
Tested with Adam optimizer with 3 qubits, but the fidelity saturated at 42% even after training for 1000 epochs for some reason. On the other hand, gradient free method that was previous used in git has 90% fidelity.

Also, worked on scaling up to 8 qubits using the same gradient free method as before and the fidelity is around 90%.